refactor: usePagination 사용 위치 변경
usePagination 훅을 불필요하게 두 번 실행하고 있어서 수정함.
