+++ src/interface/props.d.ts
@@ -1,0 +1,5 @@
+export interface IPageButtonList {
+  startPage: number;
+  endPage: number;
+  pageSize: number;
+}
